Added logic to delete Damage and Good product Quantity

+++ service/src/main/java/com/trainee/inv/service/goodquantityproduct/GoodQuantityProductServiceImpl.java
@@ -69,6 +69,10 @@ class GoodQuantityProductServiceImpl implements GoodQuantityProductService {
 
 	@Override
 	public void delete(int id) {
+		boolean existsById = goodQuantityRepository.existsById(id);
+		if(!existsById) {
+			 throw new IllegalArgumentException("ID not found");
+		}
 		goodQuantityRepository.deleteById(id);
 	}
